Move convert_eof() to ptvsd.socket.
